Move purpleair base entity to separate module
